+++ koreader-plugin/lua/_meta.lua
@@ -2,5 +2,5 @@ local _ = require("localsend_i18n").translate
 return {
     fullname = _("PDF to EPUB Receiver"),
     description = _([[Receive books from PDF to EPUB OCR over the local network.]]),
-    version = "v0.12.0",
+    version = "v0.12.1",
 }